migrated spline data to bot_utils package under namespace bot_utils
