+++ src/bot_utils/CMakeLists.txt
@@ -27,6 +27,7 @@ include_directories(
 ## Declare a C++ library
 add_library(${PROJECT_NAME}
   src/bot_utils.cpp
+  src/spline_data.cpp
 )
 
 add_dependencies(${PROJECT_NAME} ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
