+++ src/ee4308_hector/include/ee4308_hector/hector_trajectory.hpp
@@ -1,34 +1,11 @@
 #ifndef H_TRAJECTORY_HPP
 #define H_TRAJECTORY_HPP
 #include "bot_utils/bot_utils.h"
+#include "bot_utils/spline_data.h"
 #include <vector>
 #define NaN std::numeric_limits<double>::quiet_NaN()
 
-struct SplineData2D
-{
-    std::vector<bot_utils::Pos2D> spline;
-    int curr_spline_id = -1;
-    double avg_speed = NaN;
-    double target_dt = NaN;
-    int get_num_targets();
-
-    int find_pos_id(bot_utils::Pos2D &pos);
-
-};
-
-struct SplineData3D
-{
-    std::vector<bot_utils::Pos3D> spline;
-    int curr_spline_id = -1;
-    double avg_speed = NaN;
-    double target_dt = NaN;
-    int get_num_targets();
-
-    int find_pos_id(bot_utils::Pos3D &pos);
-};
-
-
-std::pair<bot_utils::Pos2D,int> get_best_goal(SplineData2D &tspline , bot_utils::Pos3D &hector_pos , bot_utils::Pos2D &turtle_pos,  double h_avg_speed);
+std::pair<bot_utils::Pos2D,int> get_best_goal(bot_utils::SplineData2D &tspline , bot_utils::Pos3D &hector_pos , bot_utils::Pos2D &turtle_pos,  double h_avg_speed);
 
 
 std::vector<bot_utils::Pos3D> LinearVert(bot_utils::Pos3D &pos_begin , bot_utils::Pos3D &pos_end);
@@ -43,7 +20,7 @@ void TrajectoryGenerationHandler(bot_utils::Pos3D current_goal ,
                                  bot_utils::Pos3D next_goal , 
                                  bot_utils::Pos3D h_pos , 
                                  bot_utils::Pos3D h_vel,
-                                 SplineData3D &tspline,
+                                 bot_utils::SplineData3D &tspline,
                                  double average_speed , 
                                  double target_dt , 
                                  double height ,
